🔨fix(dto): Integer -> Long으로 바꿈

+++ commitmood-batch/src/main/java/com/ssafy/commitmood/dto/GithubCommitDto.java
@@ -45,9 +45,9 @@ public class GithubCommitDto {
 
     @Data
     public static class StatsDto {
-        private Integer additions;
-        private Integer deletions;
-        private Integer filesChanged;
-        private Integer total;  // total_changes
+        private Long additions;
+        private Long deletions;
+        private Long filesChanged;
+        private Long total;  // total_changes
     }
 }
